+++ spec/chewy/query/nodes/match_all_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-describe Chewy::Query::Nodes::Raw do
+describe Chewy::Query::Nodes::MatchAll do
   describe '#__render__' do
     def render &block
       Chewy::Query::Filters.new(&block).__render__
     end
 
-    specify { render { r(term: {name: 'name'}) }.should == {term: {name: 'name'}} }
+    specify { render { match_all }.should == {match_all: {}} }
   end
 end