+++ spec/chewy/query/nodes/missing_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 describe Chewy::Query::Nodes::Missing do
   describe '#__render__' do
     def render &block
-      Chewy::Query::Context.new(&block).__render__
+      Chewy::Query::Filters.new(&block).__render__
     end
 
     specify { render { !name }.should == {missing: {term: 'name', existence: true, null_value: false}} }